Added scene_node model transformation chaining.
This changes means a material needs our model matrix. Ew. Nonetheless it should
work and fits the usage of the Material class, but not quite it's name. Also,
this change makes the render function of a scene node a member. It was
necessarily in one uncommitted iteration of the code, but isn't now, that being
said it's probably a good idea to keep it in so we can introduce some cache
variables as private members in the future without breaking our interface.