The Object and Material now work with textures better (and are slightly more consistent with each other).
The Object now prepares its material when preparing it's mesh (which is now
required). Also, the Mesh allocates a software texture for it's texture field.
We still have it typed as a maybe_owned because it will allow sharing textures
between materials. Same reason we are still doing it for the software mesh in
the object struct, because it will allow for the sharing of meshes, even if the
implementation doesn't change to the object. Of course, Software_Texture is
just a wrapper so it can have any sort of implementation, but the object
doesn't care about that!