Updated the Mesh interface.
Before, it was limited in the fact that vertices had to be known at preparation
time, things are now much more separate which makes it more versatile for the
client and the implementer when it comes to optimizations. It's already fully
implemented!

Also since it was a significant interface change, the functions to generate and
terrain mesh and load one from a file have been updated. We still allow for
construction via a mesh_data object so they both just use that.