Added our new event system. Simple, yet (hopefully) effective.
It uses boost signals and slots to manage all listeners/observers that wish to
be notified, (most likely) the controller will go through each element checking
if events should are enabled and manage all the mouse logic. This will be
forwarded to the element in question which will then notify all it's
dependents.