On one hand, it might be a good idea to be explicit about all indices, but eh.
I commented out that code in place of a cleaner implementation of what it has
been doing. I can't tell if they are equivalent results because blender seems
to be recalculating normals when loading in the object. It may even be better
to only write normals and uv coordinates if specifically asked for, since this
is mainly for debugging purposes anyway.