Updated the gfxlib.
The big thing is Mesh_Data now is split into a bunch of different
representations. The Mesh ABC (implemented by the Driver) is also now more
versatile, not committing every user with one specific representation. That is,
it allows for configuration of buffers with an optional element array. This
mandated a maplib refactoring/reimplementation of the Terrain => Mesh function.
The new implementation is hella slow and will probably be reworked to the old
one (while supporting the new mesh data representations). Because of the new
representations also, there really is no need for a software_mesh.

Also a bunch of graphically related things were in the graphics lib, which no
longer is the case. There is now a gfxsupportlib which has all the (optional)
goodies. It has a dependency on gfxlib.

IDriver_UI_Adapter also is *about* to be a lot more efficient, reducing a *lot*
of OpenGL state changes by just concatenating each new rectangle onto a single
mesh. The only thing is we have no way to just stop drawing the UI if it hasn't
changed.

Some tests haven't been updated and so they are disabled for now.

In addition, this commit is much to big. Oh well.