The Maybe_Owned will now prevent a move that would cause the destruction of its owned pointer unnecessarily.
If one moved from an owned maybe to another, then from that original maybe into
that other maybe a second time, that would be moving a non-owned maybe into the
maybe that actually owned the data, before this would cause the data to be
destructed and both maybe's to end up pointing to invalid memory.

This is also being tested.