The write_data_to_mesh and friends now accepts a maybe owned that it passes right into the mesh chunk it creates!
This is so that the client code doesn't have to set the mesh chunk's mesh to
own the mesh if it wants. It wasn't really a problem currently, but as more
objects are created and hidden away, it would be nice to package ownership in
the mesh chunk and not worry about it otherwise.