Fixed triangle collision + projection response.
Turns out we should be flattening the triangle before checking for collisions.
Also there is still a problem where the player stays on a flat triangle longer
when it should have detected another triangle that has an incline. The result
is the player (after a point) makes a sharp jump which isn't very pretty. I
believe this is inherent with the method of collision detection.