Implementation of event spread supporting timings!
The test program is awesome! That being said, the Event_Map interface could be
more specific as to what is allowed. For example, the test code has to allocate
both the visited positions map and the event timer map. It has to make sure
they are the same size, etc. This could be better done behind the Event_Map
constructor.