Finally clean up the Element's input mechanisms.
Instead of click, hover, and drag listeners we only use a generic on_step_mouse
callback that is sent the mouse state every time the element recieves one. To
support clicking functionality two helper callback functors can be used to only
run the callback when either the mouse is clicked or released. It's much
simpler and versatile.