Remove heightmap gen in favor of landmass gen
World / Map generation will be split up into the following steps:
 - Landmass generation
 - Heightmap generation
 - Biome and climate determination
 - River generation
 - Tree / foliage generation
Lua bindings can be written to alter each step which will be neat. Of course
we'll need to add some code to handle LOD of the terrain. After that we'll need
to figure out reasonable water rendering (etc) and it'll be good.