New rough draft for the mod interface
There is a module that the package.lua must include called redcrane for now.
That returns a table containing for now: a single register function taking a
function and a configuration object. The code is also a bit cleaner it will
hopefully be easier and clearer to clean up and generalize.