Add new class to manage a list of registered mods
It seems to work but I haven't tested it much. Basically, at initialization
time, mods are added to a vector in a declarative way (parameters types are
specified, etc). This list of mods maintains the list of registered mods and
provides a way to unregister them. It should be able to be used to add
different entry points for the lua code. Actual usage of any callbacks has not
been implemented.