initialize_vertices: Clarify code and remove gaps
The current usage of x and y will produce large gaps where each (small) square
starts at 0.0, then 1.0, then 2.0, etc. Also it's not every level that gets it
own mesh data it is per node, which is reflected in the loop but not in the
comment that is now better.