terrainlib: Physical size is now a volume
I changed the name of the other volume to avoid confusion. Chunk::vol is now
uv_vol because it represents the volume of the heightmap. Chunk::physical_size
is now a volume and it's called world_vol. The function set_physical_size and
its tests have been merged with the function and tests of set_volumes
respectively.