Search for correct asset directory with config
We now support running from the install directory. Common directories are
searched for the config file at which point it is considered the correct asset
directory. Currently this is the directory we switch to at the start of the
program, later on we really should cache the result and use it more often,
potentially. Maybe that would be stupid, I kind of like how it works now where
we don't have to worry about any particular directory, knowing things are
relative to all of our assets.