Survival lua code sorta works now
Required massive redesign of the whole thing because I found out I can't mix
Lua/C with LuaJIT ffi. Basically, I was passing around the engine as a
lightuserdata but interpreting it as luajit cdata.

This commit includes a bunch of changes in redcrane.cpp and architecture
changes in survival/.

Things that still need to be cleaned up:
 - Startup configuration (see redcrane.cpp:redc_init_engine)
 - Internal (lua) configuration (see survival/config.lua)
 - Scene management / object abstraction