Refactor Object / Camera handling in lua interface
We are now using ids and passing those around to reference objects, this means
we could technically have a vector dynamically resize, but right now we use a
pre-allocated array of object tagged-unions. The object can either be a camera
or mesh and every object has a parent (which is actually a pointer, maybe we
should use an id, that pointer is what prevents us from going
dynamically-allocated).