Add initial work on peer ptr
A peer pointer will allow both Lua and the engine to manage a resource that
doesn't make sense to have kept around if either of them aren't using it. For
example the engine can uninitialize a mesh when the engine itself is
initialized even if Lua still has a reference to it. This applies in the
reverse direction. Probably, I will still need to figure out what to do when I
need a temporary reference to some resource, etc.