Make Peer_Ptr more versatile
Client code can now construct a peer pointer from a unique pointer and create
new peers from that one instance. This should simplify code in the common case
but also allow for complex use like the make_peer_array function that fills a
std::array with a given amount of peers.

This functionality is partially tested.