Implement a basic (lua) sandbox
The fact that we had to run some engine lua code without a sandbox called for
the separation. Now, all the internal lua/ffi binding code is compiled in
advance into the executable and added to package.preload at runtime. The mod
code can require this code but it currently can't load any other code (from the
filesystem). The sandbox is also basically non-functional, but that should be
fairly easy enough to fix because of the new sandbox.lua that returns the table.