Implement sandboxing
Lots of refactoring as far as design and control flow goes.

- The engine is now constructed in two parts. First client code must use
redc_init_engine and then use redc_init_client or redc_init_server to
initialize that subsystem of the engine, respectively. redc_init_server is
neither implemented nor part of our interface as of now. This also means the
engine struct itself is just configuration, the share path, other shared
members, and an optional client. Again, the server struct doesn't even exist
yet.
- Configuration options are now universally accessed and set through lua where
each option may individually be either runtime or non-runtime (see
src/lua/init_engine.lua). Runtime options are made such that setting them and
accessing them is done through a given setter / getter pair of functions. As of
now we are not actually using this mechanism for anything.
- We are delegating more to un-sandboxed lua code. First ffi functions are
loaded putting "ffi" into package.preload. Then the configuration file is run
in a sandbox without engine functions. The config is passed to an init
procedure (also written in lua) that checks the config and adds all the bells
and whistles; this allows us to pass it to the C code as a Redc_Config struct
type without worrying that LuaJIT will fail to convert it or we get weird /
malicious data.
- Delegation of server modes is now going to be done in lua code because it
will have access to the engine and the config so it can just load the config
and server entry files and run them under a sandbox we pass as a parameter.