Add start of map functionality
Currently a map is just defined as a model (.obj file), but we'll eventually
expand this with more information. As of now, maps are loaded into the
top-level engine structure because maps are going to be needed by both the
server and the client. Every map will always have a collision shape (generated
on-load) and a mesh for rendering (which includes normals, uvs, etc). Right now
the client maintains a cache of the mesh_chunk for the active map. Old mesh
chunks for maps that are no longer active will probably be discarded (it uses
the Cache template). We can implement more complicated management in the
future. As of now the map is rendered by physics are ignored, even if it just
means adding the collision shape to the world along with the character
controller (I'm hoping it is that simple).