Add player creation mechanism that works alongside scenes
So far so good, server is being accessed through an event-based interface so
hopefully it will be easy enough to port to over the network if it comes to
that. Furthermore, we are using kind of a hack where cameras are flagged to
follow the active player but the active player is itself decoupled from the
server side so that's a win. Now it's time to fix up the player controller so
the player can move around, etc.