Implement map loading in terms of engine events.
This is in contrast to how events used to be handled in the server instance,
being polled by the scene step routine. At this point, the map is loaded, an
event is queued and the client and server (and scenes) get to execute their own
code, optionally filling in sub-components of the map. At this point the Map
structure has a Rendering_Component and a Physics_Component which are filled in
by the client and server handler, respectively. This works pretty well but the
Physics_Component could probably be generalized to work with all objects. In
fact, the two components are so new and specific to the Map that for now I
wouldn't even call it a component system, although their use seems pretty
promising.