Make a half-decent character controller
Uses a player state (either grounded or jumping) and supports actual
jumping (via spacebar, etc). Uses a ray trace to figure out if the
player is on the ground and changes things a bit to keep everything
stable. This solution is okay, but the variable shoe_size has to be
pretty large for it to actually remain stable, otherwise the character
jarringly toggles between jumping and grounded during the penetration
recovery. This obviously throws off the camera, etc.