Use the raytracing result more effectively
Instead of snapping the player to the ray collision point, let it be.
This will stop any instability caused by a snap that either causes
further intersection with the map or something weird. At the moment it's
a lot simpler it just offloads all that penetration recovery to that
segment of code, which was pretty much handling it fine anyway. The two
things were just getting in the other's way.