Change the way shaders work and are loaded
String "tags" are used to identify glsl uniforms. They are standardized
so they fill the need for functions that work on any shader that set
common properties such as projection, view and model matrix. New tags
can also be created so it's (hopefully) an easier to maintain solution
compared to caching location with set_*_name and having a method like
set_model and set_view for each new uniform we want to add.

Furthermore, there's been a big change in how shaders are created in
terms of the C interface. Lua, which directs the whole process, now puts
the shader file (<shader_name>.lua) in a sandbox where only
shader-related functions are available. The functions append strings to
shader sources until everything is done and they are compiled. The
shader file is completely declarative and acts like a DSL. It also
properly constructs tags in the new shader, so as long as everything is
done through this shader file, acting on each shader generically in C++
and Lua code should be fairly easy. Notice in the updates to C++ code
that when shaders are initialized all variables must be tagged, this is
different from how it used it work, but now locations don't need to
cached locally, which can become a pain. Tagging a variable caches its
location in a hash map, but most importantly it gives a standardized
name for it if applicable. The one issue with this approach is that it
is not trivial (nor implemented at the moment) to load a shader DSL file
from C++ without running Lua code to do it specifically.

Also if the shader DSL file cannot be found, vs.glsl and fs.glsl are
loaded instead from a directory given by the shader name (which would
have been the filename of the DSL file). This logic can be found in
src/lua/shader.lua.

Shaders must be explicitly linked at this point and the load_*_part
functions of gfx::Shader now accept source code as a string, instead of
a filename. There are now helper functions to load each shader type from
a file, but these are not members of the Shader class. The one thing
that hasn't changed is the fact that the function replaces whatever
shader source was last passed to the function.

Also stop compiling the location cache because it wasn't being used and
it isn't really relevant.