Support player crouching
This change required some changes to player measurements and how they
were calculated. The new struct Player_Dimensions is used for values
that may change when the player crouches. This prompted better
practices in using the movement speed, which should be defined in metric
units. In fact before I believe it was technically frame-rate dependent
if bullet wasn't using fixed-time steps itself. In other words, it
didn't factor in dt.