Set the OpenGL viewport to the window size
This is sort of a workaround for the fact that SDL creates a context
with the width and height equal to the parameters of SDL_CreateWindow.
If passing a zero width and height for fullscreen mode, the opengl
context won't work (or is more probably just really small), but this
fixes that.