Handle bad semantics by returning a bad bind
Instead of crashing we just return a bad bind. OpenGL doesn't seem to
complain and now it's easy enough to check before making an OpenGL call.

This commit adds an Attrib_Semantic to String function that could be
used to log the issue when a bad bind pops up. That being said, we are
in a render routine so any logging will effectively spam the user so we
can't actually log this until we find a better solution.