Add sized texture formats to Texture_Format enum
Useful for when we want some specific internal format. That being said,
it might be better in the future to somehow mix and match the format and
type with two enums and create the GL enum on the fly, similar to how
Upload_Hint and Usage_Hint work, but I don't know how to make this not a
bitch-and-a-half to actually implement correctly (and properly change a
lot of interfaces).