Document runtime (IDriver) vs static (funcs.h) abstractions
They both abstract OpenGL currently, but it's probably best to move in
the direction of the runtime abstraction when it comes time to test
Vulkan and DirectX on Windows. It's most likely true that the calls to
the graphics API will be much more of a bottleneck then any virtual
function, dynamic type checking, etc will ever be.