Refactor graphics system
glTF loading code now uses the Driver / runtime / object-oriented
interface. The reason this had to be done was because gfx/scene.cpp was
basically using a bunch of functions that couldn't be changed at
runtime. That way the engine couldn't use both DirectX and OpenGL, we
would have to pick one at compile time. Although it's not the best, the
interface is a lot better now. Buffers and meshes are now separate
concepts, which is important because glTF makes use of shared buffers
between meshes a lot. A lot of gfxsupportlib (now gfxextralib) code was
updated as well and it seems a little bit nicer, slightly more
consistent. Asset loading / caching still need a little bit of work.
Asynchronous loading would be nice.