Deferred_Shading: Use an array of lights instead of a vector
In other words, use a pointer to a light and a length. There is no
reason to mandate vector storage. I suppose if we want to get really
wacky we could use iterators or something. I've never really thought
about this kind of approach for these kinds of things. TODO amirite?