Refactor glTF loading code
Load meshes like nodes used to load and vice versa. This change is
important because it makes it easier when we want to extend the
load_nodes function to add both regular nodes and light nodes. The two
nodes will go in separate vectors, which will be hard to implement the
way things worked before. Now, since meshes probably won't change in the
future (that's a good one), it will just have to use this rigid loading
scheme. The reason we do this in the first place at all is because there
is a circular dependency when it comes to techniques / materials
possibly using transformations from nodes.