Improve fake mesh code by using numpy
Instead of using an object name and a mesh name, we look up the mesh
name by going through the node, which we need for vectors, and parents,
etc anyway. Matrix multiplication is also a lot simpler now with numpy,
however it's a little funky because numpy expects row major matrices.
Instead of worrying about it we just do pre multiplication instead of
post multiplication, etc.