Toggle desk lamp in the office
This changes the way maps are managed in Lua. Basically, Lua needs a
reference to the map so it waits for a new kind of Lua event to give the
signal that the map has been loaded, it includes a cdata/void pointer to
the map that is stored and used later for setting light state.

Events sent to lua are now literally called Lua events. There is one
weird thing about how the map loaded event propagates. It is sent by the
server event callback on a map loaded event (which itself is raised by
the engine). There is a question about whether asking for server events
makes sense in the client code. I mean right now there isn't much of a
distinction, there is just some confusion.