Improve blender/gltf/engine light compatibility
The state of the light (on or off) can be set in the glTF, and
fallOffAngle is actually defined as the diameter, not radius, so we do
the conversion on load and forget about it. Intensity is read from the
energy field and comes directly from blender. We also have a few more
lights in the blender scene that need to be turned off when the game
starts, so we added a mechanism to add that in convert-scrips/office.py.