Initial generator control code
Supports multiple levels that cause the generator to become more and
more unstable, causing jumps in the effective load of the generator. We
tie this to the lights and they get dimmer and dimmer and tend to
flicker, although the noise function isn't really ideal in this case,
since it rarely causes the light to completely shut off.