Simplify mod config loading
The ability to use functions for querying the value of dynamic options at
runtime was cool and all but it wasn't even used. Removing that functionality
allows us to remove most of the functionality in the metatable. Now it is only a
simple one that keeps the configuration read-only once its created.