Remove persistent state tracking in asset_render
Rendering_State made sense for small demos, but when state changes are being
applied by things that are not asset_render, it becomes a hindrance.

Now, driver and deferred shader helper are being provided in parameters, while
a simplified Rendering_State struct is used to keep track of state for the
duration of the function call.